# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Agent System",
    "description": "\n    ## Financial Agent API\n    \n    This API provides financial analysis capabilities powered by AI agents.\n    \n    ### Features:\n    - Multi-agent financial analysis\n    - Invoice and contract processing\n    - Revenue analysis\n    - Interactive visualizations\n    - Session management\n    ",
    "contact": {
      "name": "Financial Agent API Support",
      "email": "support@financialagent.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/{user_id}/sessions/{session_id}/download-data": {
      "get": {
        "tags": [
          "📥 Download"
        ],
        "summary": "Download Session Data",
        "description": "Download session analysis data as CSV",
        "operationId": "download_session_data_api_users__user_id__sessions__session_id__download_data_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visualization-json": {
      "get": {
        "tags": [
          "📥 Download"
        ],
        "summary": "Get Visualization Json",
        "description": "Proxy endpoint to fetch plotly JSON from blob storage\nThis avoids CORS issues by proxying through our backend",
        "operationId": "get_visualization_json_api_visualization_json_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/csv-data": {
      "get": {
        "tags": [
          "📥 Download"
        ],
        "summary": "Get Csv Data",
        "description": "Proxy endpoint to fetch CSV data from blob storage\nThis avoids CORS issues by proxying through our backend\nReturns parsed JSON array for frontend table display",
        "operationId": "get_csv_data_api_csv_data_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}